+++ src/main/java/com/git/bookstore/util/TokenUtility.java
@@ -14,7 +14,7 @@ public class TokenUtility {
     private static final String TOKEN_SECRET = "Signature";
 
     //use to create new token
-    public static String createToken(Long userId) {
+    public static String createToken(int userId) {
         try {
             //to set Algorithm
             Algorithm algorithm = Algorithm.HMAC256(TOKEN_SECRET);
@@ -29,7 +29,7 @@ public class TokenUtility {
     }
 
     //use to decode the token
-    public Long decodeToken(String token) {
+    public int decodeToken(String token) {
         int userId;
         //for verification algorithm
         Verification verification = null;
@@ -44,7 +44,7 @@ public class TokenUtility {
         DecodedJWT decodedJWT = jwtVerifier.verify(token);
         Claim claim = decodedJWT.getClaim("user_id");
         userId = claim.asInt();
-        return Long.valueOf(userId);
+        return userId;
     }
 
 }
